Added cat command.
